% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apfd.R
\name{apfd}
\alias{apfd}
\title{Annual Proportional Flow Deviation}
\usage{
apfd(flow, dates, ref_start, ref_end)
}
\arguments{
\item{flow}{vector of (monthly) streamflow data}

\item{dates}{vector of same length as flow containing the dates associated with the flow values
the format is expected to be interpretable by lubridate::parse_date_time to convert to date format}

\item{ref_start}{starting date of the reference period; has to be in the same format as date_format}

\item{ref_end}{end date of the reference period; has to be in the same format as date_format}

\item{date_format}{character vector defining the date-time format of dates see lubridate::parse_date_time for details}
}
\value{
numeric value of the flow alteration ranging from 0 (unmodified) to 3.46 (fully modified)
}
\description{
computes the annual proportional flow deviation of monthly stream flow data
according to Gehrke et al. (1995)
}
\references{
Gehrke et al. (1995):River regulation and fish communities in the murray‐darling river system,
Australia Regulated Rivers: Res. Manage. 11 363–75,
https://doi.org/10.1002/rrr.3450110310
}
\author{
Florian Betz
}
