% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Estimating annual volume from longterm discharge or vice versa}
\usage{
convert(discharge, volume, conversion = "to_volume")
}
\arguments{
\item{discharge}{numeric value of the discharge in m3/s, only needed for estimating volume (the default)}

\item{volume}{numeric value of the annual water volume in km3, only needed for estimating annual average discharge}

\item{conversion}{either "to_volume" or "to_discharge", giving the direction of calculation}
}
\value{
numeric value giving the annual water volume or the annual average discharge
}
\description{
provides an estimate of annual water volume from average discharge or vice versa
This requires some caution as it is a rough estimation only when calculated based on average annual flow
}
\note{
arguments need to specified explicitly, otherwise wrong estimations might be the consequence.
}
\author{
Florian Betz
}
