% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dor.R
\name{dor}
\alias{dor}
\title{Degree of Regulation}
\usage{
dor(cap, flow)
}
\arguments{
\item{cap}{reservoir capacity in cubic kilometer (km3)}

\item{flow}{annual inflow into the reservoir (km3)}
}
\value{
numeric value with the degree of regulation
}
\description{
computes the degree of regulation (Lehner et al. 2011) using
reservoir capacities and annual streamflow as input
}
\references{
Lehner et al. (2011): High-resolution mapping of the world's reservoirs
and dams for sustainable river-flow management Front. Ecol. Environ. 9 494–502,
https://doi.org/10.1890/100125
}
\author{
Florian Betz
}
