% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_volumes.R
\name{estimate_volume}
\alias{estimate_volume}
\title{Estimate yearly water volumes from (monthly) discharge time series}
\usage{
estimate_volume(discharge, date, date_format)
}
\arguments{
\item{discharge}{vector with (monthly) discharge values in m3/s}

\item{date}{character vector indicating the dates associated with discharge}

\item{date_format}{character defining the format of the date vector. See lubridate::parse_date_time for details}
}
\value{
data.frame of two colums, the year and the associated volume estimation
}
\description{
This function allows to estimate the annual volume at a gauge from the discharge time series. Currently, this function
assumes monthly discharge time series, but daily time series will be supported soon.
}
\author{
Florian Betz
}
