% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_neespi.R
\name{read_neespi}
\alias{read_neespi}
\title{Function to read data from the NEESPI database}
\usage{
read_neespi(file)
}
\arguments{
\item{file}{path to the csv file}
}
\value{
data.frame with three columns: "date" in format "Date" containing the date; "month": the month (e.g. for hydrograph computation); value": the actual value of the data series
}
\description{
This function reads the csv files of the Central Asian hydro and climate sites downloaded from https://neespi.sr.unh.edu/maps and
converts them into a data.frame with all required information for further processing
}
\author{
Florian Betz
}
