% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrappingEfficiency.R
\name{trappingEfficiency}
\alias{trappingEfficiency}
\title{Compute sediment trapping efficiency of reservoirs}
\usage{
trappingEfficiency(
  capacity,
  discharge,
  drain_area,
  is_m3_s = FALSE,
  method = c("brune", "brown", "dendy")
)
}
\arguments{
\item{capacity}{storage capacity of the reservoir in cubic kilometer (km3)}

\item{discharge}{annual discharge at the reservoir in km3;
if given as annual average in m3 per second, it can be converted by setting is_m3_s to TRUE}

\item{drain_area}{drainage area of the catchment of the reservoir in m2; only used if method="brown"}

\item{is_m3_s}{logical; if FALSE, the function assumes discharge to be in km3;
if TRUE, then discharge is assumend to be in m3 per second (e.g. annual average) and converted to km3}

\item{method}{which method to use for the computation; can be "brune", "brown" or "dendy"}
}
\value{
numeric value giving the trapping efficiency from 0 (no sediment trapped) to 1
(potentially all sediment is trapped)
}
\description{
Compute sediment trapping efficiency of reservoirs
}
\references{
tba
}
\author{
Florian Betz
}
